% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yieldfns.R
\name{calcyield}
\alias{calcyield}
\title{Estimate HIV testing yield under a number of strategies}
\usage{
calcyield(data, y, strategies, weight = "No Weights",
  cluster = "No Clusters", stratlabels = NULL)
}
\arguments{
\item{data}{Input dataset}

\item{y}{indicator that HIV test is positive}

\item{weight}{name of weight variable, defaults to "No Weights"}

\item{cluster}{name of variable indicating clusters, defaults to "No Clusters"}

\item{stratlabels}{optional vector of strategy labels to improve readability}

\item{strategyind}{name of variable indicating whether or not a specific person would be tested under a given strategy}

\item{strata}{name of variable to be stratified by (defaults to NULL)}
}
\description{
This function summarizes results across multiple strategies
}
\examples{
calcyield(data = mydata, y = "hiv", strategies = c("str1", "str2", "str3"), stratlabels = c("All venues", "Some other strategy", "Another one"), weight="weight", cluster = "sitespotid")
}
