% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yieldfns.R
\name{plotyield}
\alias{plotyield}
\title{A function to plot expeceted HIV testing yield by strategy}
\usage{
plotyield(results)
}
\arguments{
\item{results}{Matrix of results from calcyield function}
}
\description{
This function plots the number needed to test to identify 1 new (previously undiagnosed) case of HIV and 95% CI for each strategy
}
\examples{
plotyield(allresults)

}
