% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yieldfns.R
\name{yieldstrategy}
\alias{yieldstrategy}
\title{A function to estimate yield and NNT from each strategy}
\usage{
yieldstrategy(data, y, strategyind, strata = NULL,
  weight = "No Weights", cluster = "No Clusters")
}
\arguments{
\item{data}{Input dataset with one record per individual not previously diagnosed with HIV}

\item{y}{name of variable with indicator that HIV test is positive}

\item{strategyind}{name of variable  indicating whether or not a specific person would be tested under a given strategy}

\item{strata}{name of variable with values of variable to be stratified by (defaults to NULL)}

\item{weight}{name of variable with respondent weights  (e.g., survey sampling weight), if applicable (defaults to NA)}

\item{cluster}{name of variable with cluster IDs, if applicable (defaults to NULL)}
}
\description{
This function computes HIV testing yield for a specific strategy
}
\examples{
yieldstrategy(data = mydata, y = "hiv", strategyind = "s1", strata = "sex", weight = "personweight", cluster = "venue")
}
